Moved the apply button to its own component
